requirements.txt
requests>=2.31.0
google-generativeai>=0.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0
click>=8.0.0
prompt-toolkit>=3.0.0
markdown>=3.4.0
pygments>=2.15.0
colorama>=0.4.6 